+++ qbraid/transpiler/conversions/qasm2/qasm2_to_ionq.py
@@ -21,12 +21,12 @@ import openqasm3.ast
 
 from qbraid.passes.qasm3.compat import convert_qasm_pi_to_decimal
 from qbraid.programs.circuits.qasm import OpenQasm2Program
-from qbraid.programs.typer import IonQDict, Qasm2String
+from qbraid.programs.typer import IonQDictType, Qasm2String
 from qbraid.transpiler.annotations import weight
 
 
 @weight(1)
-def qasm2_to_ionq(qasm: Qasm2String) -> IonQDict:
+def qasm2_to_ionq(qasm: Qasm2String) -> IonQDictType:
     """Returns an IonQ JSON format representation the input OpenQASM 2 string.
 
     Args: